fix(mql5): unblock MT5 compilation + static validator guard
Fixes the "EA refuses to compile in MetaEditor" crisis. All 11
templates now produce MQL5 that MT5 accepts, and a post-pass
validator catches regressions before any future export ships.

── Root-cause fixes in the codegen

1. isLong / slPriceDistancePips were undeclared.
   Every stopBody translator (exit.atrBased, risk.atrRisk, exit.rrBased,
   …) references `isLong` and `slPriceDistancePips` — but the template
   that wraps stopBody never declared them. Result: "undeclared
   identifier" errors for the majority of templates.
   Fix: declare them at the top of ZxOpen() in template.ts alongside
   entryPrice/slPrice/tpPrice.

2. AccountInfoDouble was called with 2 arguments.
   batch2-exec.ts used AccountInfoDouble(ACCOUNT_MARGIN_LEVEL, ml) —
   that overload doesn't exist in MQL5. Rewrote MarginLevelPct() to
   use the single-arg form with a guard for zero margin.

3. r.position = PositionGetInteger(POSITION_TICKET) lost data.
   r.position is ulong; PositionGetInteger returns long. MetaEditor
   flags this as a warning under #property strict. Added explicit
   (ulong) casts in manage-trailingStopAtr.ts and manage-partialClose.ts.

4. TimeHour / TimeMinute / TimeDayOfWeek are deprecated.
   MT5 warns on them under strict and may remove them. Replaced 8
   occurrences across 6 translators with new helpers in template.ts
   (ZxHour / ZxMinute / ZxDayOfWeek / ZxDayOfYear / ZxMonth) that read
   TimeToStruct(TimeTradeServer()) — the documented replacement.

── Prevention: lib/mql5/validator.ts

Post-pass that runs after assembly on every compileStrategy() call.
Diagnostics flow back through CompileResult.diagnostics so the UI can
surface them alongside graph errors. Checks:
  · unbalanced braces / parens (error)
  · deprecated Time* functions (warning)
  · AccountInfoDouble 2-arg signature (error)
  · r.position without ulong cast (warning)
  · duplicate top-level function definitions — catches the case
    where combining two risk blocks emits RiskPercentPerTrade() twice
    (error)
  · ZxOpen-only identifiers (isLong, entryPrice, slPriceDistancePips)
    appearing outside ZxOpen — catches the exact class of bug #1
    regressing (error)

The string/comment-aware stripper prevents false positives on text
inside Print/StringFormat/etc.

── Tooling

scripts/gen-sample.ts — dump one template's MQL5 to /tmp/zx for
manual inspection.
scripts/gen-all.ts — generate all 11 canonical templates; prints
diagnostics per template.
scripts/check-validator.ts — smoke-test every detector against a
known-bad fragment AND confirm every canonical template passes clean.
Run with `npx tsx scripts/check-validator.ts`.

All 11 canonical templates: 0 errors, 0 warnings after the fixes.

+++ lib/mql5/translators/manage-partialClose.ts
@@ -30,7 +30,7 @@ export const translate_manage_partialClose: Translator = (node) => {
       if(closeVol > 0) {
         MqlTradeRequest r; MqlTradeResult rr; ZeroMemory(r); ZeroMemory(rr);
         r.action = TRADE_ACTION_DEAL;
-        r.position = PositionGetInteger(POSITION_TICKET);
+        r.position = (ulong)PositionGetInteger(POSITION_TICKET);
         r.symbol = _Symbol;
         r.volume = closeVol;
         r.type = (dir == POSITION_TYPE_BUY) ? ORDER_TYPE_SELL : ORDER_TYPE_BUY;
